+++ src/gameIndex.js
@@ -3,5 +3,5 @@ export const gameIndex = {
   Gribbles: "gribbles",
   Thirdle: "thirdle",
   Palette: "palette",
-  TwoD: "twod",
+  Packed: "packed",
 };
