+++ src/packed/logic/getOffsets.js
